Optimize meshlets to 4 bytes per vertex from 16

+++ src/voxel/local_mesh.rs
@@ -9,7 +9,7 @@ pub struct LocalMesh {
 
 #[derive(Clone, Eq, Hash, PartialEq)]
 pub struct LocalVertex {
-    pub position: Vector3<u16>,
+    pub position: Vector3<u8>,
 }
 
 #[derive(Clone)]
